requirements.txt
py3270==0.3.5
pandas==3.0.2
openpyxl==3.1.5
python-dotenv==1.2.2